Add default method support to art
This commit starts the process of adding default methods and their
associated pieces to ART.

This adds full support for calling default methods using
invoke-interface and invoke-virtual on objects implementing the
interfaces. Verifier is changed to allow this when the runtime is
started with -Xexperimental:default-methods.

This also adds support for defining and calling static methods on
interface classes.

Directly calling overridden default methods using invoke-super is not
yet supported.

This adds 4 new run-tests for this functionality.

Change-Id: I35ca800d99d3329348b277789b70ceeeba6e7f03