Mark __jit_debug_register_code as a unique symbol on arm
This symbol is used to notify the debugger about JIT library load/unload
events. Because of ICF several functions with the same body put to the
same adress by the linker what causes the debugger to get notifictaion
about JIT load/unload events in several case where it isn't really happened.

This change disable ICF for the __jit_debug_register_code symbol to prevent
this behavior with keeping the symbol unique in the object file.

Change-Id: I392be25d8a07f47f8217bf693c8e5919d29cac7b