Reduce memory used by CompiledMethods.
Use LengthPrefixedArray<>s instead of SwapVector<>s to store
CompiledMethod data and get rid of the unnecessary members
of CompiledMethod to reduce dex2oat memory usage. Refactor
the deduplication from CompilerDriver to a new class.

Use HashSet<> instead of std::set<> for the DedupeSet<> to
further decrease the memory usage and improve performance.

This reduces the dex2oat memory usage when compiling boot
image on Nexus 5 (with Quick) by ~7.5MiB (6%) and also
reduces the compile time by ~2% (~1% overall dex2oat time)
when measured with userspace governor at frequency 1036800.
(The compile time improvement looks higher when we don't
restrict the frequency but the results are very random due
to thermal throttling.)

Change-Id: I974f1f5e58350de2bf487a2bca3907fa05fb80ea