Add support for generating CMOV for X86
Support generation of ConditionalSelect instructions, which on X86 will
be compiled into CMOVs, avoiding a jump sequence.

The input to the ConditionalSelect can come from an HCondition or a
boolean value.

Currently, this is only supported for X86 and X86_64.  It should be
straight forward to add other architectures.

Added a test for supported architectures.

Change-Id: Ieae5595d93ca13b4c3a7f30f0e4b8b5f1f7d7bfe
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>