ART: Add more deopt tests
Extend libarttest with the ability to turn off asserts for method
state. Use this to put asserts into test functions, but turn them
off if the test wouldn't react as we want, e.g., when we're in
interpreter mode.

Extend run-test 449 by adding asserts for the method state for the
expected deopt cases. This tests both standard behavior, as well
as single-frame deoptimization.

Bug: 21611912
Change-Id: I27fa5e3e44d8c5eab57962d144c6aec96587bf45