enum operator<<() script fails with do { } while cond;
The script assumes that a line containing a '}' and a ';' will only
and always signal the end of a scope.
Unfortunately, this also matches '} while (myCond);',
thus incorrectly thinking it has reached the end of a scope.

This can result in enums being emitted without scope resolution
(or with insufficient scope resolution), cause compile errors.

Change-Id: I8ba26b25dfbbf21e7869e9063367a5899ccd2ee3
Signed-off-by: randy.jeong <randy.jeong@samsung.com>