Class Linker: Implement default interface verification ordering.
We need to ensure that default interface verification failure is
propagated to the implementing class so that the compiler is able to
properly handle verification failures. Furthermore we need to ensure
that default interfaces are verified at the same time their
implementing classes are.

The does not add additional support for default methods to the verifier.

Bug: 24618811

Change-Id: Ia11a50a28889dbd7ca22ffa730ef443e39f3cb5b