Correct exception behavior for default methods
Default methods are defined supposed to throw an
IncompatibleClassChangeError (ICCE) when they are called and there is
no "best" implementation. Previously we would simply throw an ICCE
during class loading as soon as we noticed that this would happen if
called. This makes us wait until we actually attempt to execute the
method. Furthermore, this allows us to use other, non-conflicting,
methods on the object as normal.

Bug: 24618811

Change-Id: Id891958a81f9b3862b2ce5919636aabef7d3422e