Implementation of fast lookup table to search class_def by descriptor.
Lookup table is a hash table which built at compile time and stored into oat file.
At runtime the table is restored and used in the method DexFile::FindClassDef(const char*) to
perform fast search of the class_def_idx by class descriptor. Advantages of the lookup table over
the HashSet (runtime/base/hash_set.h) are:
1. Lookup table is built at compile time and uses read-only memory at runtime
2. Lookup table uses less memory then DexFile::Index (less by 80% for /system/framework/framework.jar on Nexus5)
3. Lookup table does less string comparisons compared with HashSet (less by 70% for zygote process on Nexus5)
The disadvantage of the lookup table is it increased boot.oat size by 0.2% on Nexus5 and
application .oat file by 0.3% in average on Nexus5.

Change-Id: I7246c1d9ad9fe81fe5c5907a4bf70396d8f9242a