Make FindInterfaceMethod search backwards
This means that FindInterfaceMethod will find the actual method that
default method searching would find. Because this returned method is
only used for (1) checking that there is a method, (2) printing the
name and signature, and (3) finding an actually runnable method
through other methods, this shouldn't make any real difference to
overall behavior. I add this since it is no slower than searching
forward and can make for slightly more informative error messages on
e.g. invocation errors. Furthermore, having it be consistent with the
interpreter interface search ordering is useful for understanding the
code.

Change-Id: Ifc24f7a8e519da8241c069bc5705458d807ebb6f