Class Linker: Make FindInterfaceMethod consistant with ClassLinker
Changed mirror::Class::FindInterfaceMethod to search iftable backwards
to be consistent with the search direction in class linker. Here the
search direction does not actually matter but it does in ClassLinker
and so we might as well be consistent.

This also makes for (potentially) slightly more informative error
messages since the found interface method is more likely to be the one
that the class linker would select.

Change-Id: Ifc24f7a8e519da8241c069bc5705458d807ebb6f