Optimizing: Determine invoke-static/-direct dispatch early.
Determine the dispatch type of invoke-static/-direct when
building the graph. This allows the inliner to pass the
"needs dex cache" check and inline more. It also allows the
code generator to avoid requesting a register location for
the ArtMethod* for kDexCachePcRelative and direct methods.

The supported dispatch check handles also situations that
the CompilerDriver currently doesn't allow. The cleanup of
the CompilerDriver and required changes to Quick will come
in a separate change.

Change-Id: I3f8e903a119949e95871d8ab0a995f4731a13a07