X86: Use memory to do array range checks
Currently, a HBoundsCheck is fed by an HArrayLength, causing a load of
the array length, followed by a register compare.

Avoid the load when we can by comparing directly with the array length
in memory.  Implement this using an arch specific pass that
replaces the length input to the HBoundsCheck with the input to the
HArrayLength.  This allows us to generate code that references the
length in memory directly.

Only do this replacement when we are the only user of the ArrayLength
and it isn't visible to the environment.  Add an extra check to ensure
that the actual index is 'int' because some code seems to be break that.

Handle the special case where the array is 'null' and where an implicit
null check can't be eliminated.

This code moves the load of the length to the slow code for the failed
check, which is what we want.

Change-Id: I9cdb183301e048234bb0ffeda940eedcf4a655bd
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>