Move .dynsym ELF section after .text
The main purpose of the dynamic section is to describe where
.text is and how large it is.  It makes more sense to write
this section once we actually know this information.

Also write string sections (.dynstr and .strtabe) before the
corresponding symbol sections because the strings take more space.

Neither of these changes is important now since we build the
sections in memory first, but it is move towards streaming of
the ELF file directly do disk.

Change-Id: I61cc9f934b031d26a3a1f6007b51315e1538e3eb