Minor fixes and cleaning of arm64 static and direct calls code.
Fixes:
The proper way to avoid the MacroAssembler to generate code before or
after an instruction is to block the pools (usually via
`vixl::BlockPoolsScope`). Here we can use
`vixl::SingleEmissionCheckScope`, that checks we generate only one
instruction and also blocks the pools.
In practice the current code would have worked fine because VIXL would
not have generated anything after `Bl()` or `Ldr()`, but that was not
guaranteed.

Cleaning:
- `XRegisterFrom()` returns an X register. Calling `.X()` is not
  required.
- Since we are sure (after the previous fixes) that nothing will be
  emitted around the instructions we care about, update the code to
  bind labels before the instructions for simplicity.