Minor fixes and cleaning of arm64 static and direct calls code.
Fixes:
The proper way to avoid the MacroAssembler to generate code before or
after an instruction is to block the pools (usually via
`vixl::BlockPoolsScope`).  Here since we only care about one
instruction using `vixl::SingleEmissionCheckScope` also blocks the
pools and checks only one instruction is generated.
In practice the current would have worked fine because VIXL would not
have generated anything after `Bl()` or `Ldr()`, but that was not
guaranteed.

Cleaning:
- `XRegisterFrom()` returns an X register. Calling `.X()` is not
  required.
- Since we are sure (after the previous fixes) that nothing will be
  emitted around the instructions we care about, update the code to
  bind labels before the instructions for simplicity.