[WIP] Fix ClinitCheck pruning.
Make sure we merge the CllnitCheck only with LoadClass that
is a part of the very same dex instruction and do the same
checks for merging LoadClass into static invoke. This fixes
incorrect stack traces from class class initializers (wrong
dex pcs).

Remove the restriction on merging with inlined invoke static
as this is not necessary anymore. This was a workaround for
    X.test():
       invoke-static C.foo() [1]
    C.foo():
       invoke-static C.bar() [2]
After inlining and GVN we have
    X.test():
       LoadClass C (from [1])
       ClinitCheck C (from [1], to be merged to LoadClass)
       InvokeStaticOrDirect C.bar() (from [2])
and the LoadClass must not be merged into the invoke as this
would cause the resolution trampoline to see an inlined
frame from the not-yet-loaded class C during the stack walk
and try to load the class. However, we're not allowed to
load new classes at that point, so an attempt to do so leads
to an assertion failure. With this CL, LoadClass is not
merged when it comes from a different instruction, so we can
guarantee that all inlined frames seen by the stack walk in
the resolution trampoline belong to already loaded classes.

TODO: If we cannot merge a ClinitCheck with LoadClass, we
should try and merge it with a static invoke:
    X.test():
        const-class C
        ivoke-static C.foo()
After inlining:
        LoadClass C (from const-class)
        ClinitCheck C (from invoke-static)
        InvokeStaticOrDirect (from invoke-static)
ClninCheck cannot be merged with LoadClass but it can be
merged with InvokeStaticOrDirect.

Change-Id: I2b8da8d4f295355dce17141f0fab2dace126684d