ART: Register allocator improvements
Instead of spilling eagerly we try to move spill outside the loop
if possible.

Inverval is splitted inside loop exit block if we have a chance to
do it, instead of splitting between basic blocks.

In some cases we split live intervals inside loop. If we don't
have uses that requires registers after loop we having situation
when value is in register at loop start and on stack at loop back
branch. In this case not having uses on register inside the loop,
leads to excess fills on back branches.

Change-Id: Ib18e30202f1516bc5343060ad6f018447224300d
Signed-off-by: Anton Shamin <anton.shamin@intel.com>