Added the ability to specify the dex_location in dex2oat for dex files from the classpath.
Currently when preoptimization on a host is being performed there is no way
to specify the correct dex location for the dex files passed in dex2oat through
the argument --runtime-arg -cp. As the result paths in the key-value storage of
the oat file are wrong.
This change adds the option --class-path-location which we can use to specify
the dex locations separated by ':' for the files from the classpath.
How we plan to use this feature:
There are situations when a built-in application uses an external library from
the /system/framework/ but the library is not in bootclasspath. During compilation
of such applications not all classes are verified. Verification of these classes
is delayed to run time. If we pass this library to the classpath of dex2oat during
compilation of an application then classes using the external library will be
verified and we improve the application's launch time and performance.

Change-Id: Ifaac9b5d2849a48d1f1462ed19216ff5444964ed