Remove call to log function before InitLogging() is called
Using LOG() is not possible (and leads to a segfault) in the preamble
of JNI_CreateJavaVM(), since the logging facilities are initialized
afterwards when Runtime::Create() calls InitLogging().

Additionally, to prevent a segfault from happening in LogMessage's
destructor if called before InitLogging(), add a debug assertion on
the logging lock.

Change-Id: Ia21e5f37874b909b0a96c2db5683a5e45672943e