[WIP] Write dex files to oat file early.
TODO:
  - reopen dex files from oat file,
    for boot image, pass them to Runtime::Create().
  - rewrite ImageTest and OatTest to use the correct
    OatWriter::WriteDexFiles(), drop the old overload.

The oat file format has changes slightly, the class offset
table has been moved from the OatDexFile to its own section.
This actually fixes the alignment of these offsets as they
could have been unaligned previously, yet accessed as normal
with significant performance impact if the kernel has to
emulate the unaligned access (say, mips).

Note: The "image-location" stored in the OatHeader is now
the location specified through the --boot-image option (or
the default) instead of the actual filename determined by
ImageSpace::Create(). This is the same as returned by
GetImageLocation() instead of that returned by GetName().

Change-Id: I0f4799bb1f1ca28e3533156a3494f55345c3e10a