Write dex files to oat file early.
Write dex files to oat file before we actually open and
verify them. Instead, open and verify the copies from the
oat file and use these. This way, in the most common case
of zipped dex files, we have mmapped dex files instead of
inflated dex files. That reduces the number of dirty pages
used by dex2oat.

Reading /proc/self/statm after we write the oat file for
a compilation of a certain large app on Nexus 5 AOSP build
with -j1, three attempts before and after this CL gave
    before: 346061 189462 6269 26 0 140723 0
            346189 189450 6269 26 0 140851 0
            346061 189463 6269 26 0 140723 0
    after:  346186 185808 23040 27 0 140468 0
            346186 185819 23040 27 0 140468 0
            346186 185822 23040 27 0 140468 0
These values are in pages (4KiB), so while the "size"
(=VmSize) is essentially unchanged, the "resident" (=VmRSS)
is over 14MiB less and the "shared" (i.e. backed by a file)
is 65.5MiB more. That is, the amount of dirty non-pageable
memory used is reduced by about 80MiB.

The oat file format has changed slightly, the class offset
table has been moved from the OatDexFile to its own section.
This actually fixes the alignment of these offsets as they
could have been unaligned previously, yet accessed as normal
with significant performance impact if the kernel has to
emulate the unaligned access (say, mips).

Note: The "image-location" stored in the OatHeader is now
the location specified through the --boot-image option (or
the default) instead of the actual filename determined by
ImageSpace::Create(). This is the same as returned by
GetImageLocation() instead of that returned by GetName().

Change-Id: I0f4799bb1f1ca28e3533156a3494f55345c3e10a