Pass down to the runtime the application code paths.
When registering the app with the runtime the framework needs to pass
now the list of application code paths. This will be used by JIT to know
what profile info to persist.

This fixes the reliance on OatFileManager::GetPrimaryOatFile which may
produce inconsistent results based on external factors (i.e. class path
order or failing to compiling the first dex file from the class path)

Bug: 26080105
Change-Id: Iadcebd2684fcd48569e8f76ef21bd4d117fedc05